// src/types/dtos/employeeDto.ts
export type EmployeeDto = {
  id?: number;
  names?: string;
  lastnames?: string;
  email?: string;
  identification?: string;
  birthdate?: string;
  address?: string;
  phone?: string;
  vaccinationstate?: number;
  vaccine?: number;
  vaccinationdate?: string;
  dosisnumber?: number;
  username?: string;
  password?: string;
};